requirements.txt
Django==1.8.17
psycopg2==2.5.3 
django-suit==0.2.23
djangorestframework==3.3.1 
celery==3.1.20
beautifulsoup4
requests
